+++ internal/consensus/adaptor/inbound_byhash_test.go
@@ -50,7 +50,7 @@ func TestHandleNodeReply_AcceptsByHashStateNodes(t *testing.T) {
 	armFetchAcquisition(r) // a reply is only processed while an acquisition is in flight
 	r.handleFetchPackReply(&peermanagement.InboundMessage{
 		PeerID:  peermanagement.PeerID(5),
-		Type:    uint16(message.TypeGetObjects),
+		Type:    message.TypeGetObjects,
 		Payload: payload,
 	})
